+++ CHANGELOG.md
@@ -7,9 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [4.12.4] - 2026-05-06
+
+- Upgrade to SQLite3 Multiple Ciphers version 2.3.4 (SQLite version 3.53.1)
+
 ## [4.12.3] - 2026-04-13
 
-- Upgrade to SQLite3 Multiple Ciphers version 2.3.2 (SQLite version 3.51.3)
+- Upgrade to SQLite3 Multiple Ciphers version 2.3.3 (SQLite version 3.53.0)
 - Use internally new API function `sqlite3mc_cipher_name_copy` instead of `sqlite3mc_cipher_name`
 
 ## [4.12.2] - 2026-03-19
@@ -738,7 +742,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - First public release
 
-[Unreleased]: ../../compare/v4.12.3...HEAD
+[Unreleased]: ../../compare/v4.12.4...HEAD
+[4.12.4]: ../../compare/v4.12.3...v4.12.4
 [4.12.3]: ../../compare/v4.12.2...v4.12.3
 [4.12.2]: ../../compare/v4.12.1...v4.12.2
 [4.12.1]: ../../compare/v4.12.0...v4.12.1